fix: more cleanups
